removed decimal

+++ src/models/product.py
@@ -4,6 +4,6 @@ from decimal import Decimal
 class Product(BaseModel):
     name: str
     unit: str
-    cost_per_unit: Decimal
-    price_per_unit: Decimal
+    cost_per_unit: float 
+    price_per_unit: float
     quantity_in_stock: float
